refactor: move validation and query code out of login and signup route dirs
