--- www/app/routes/login/queries.ts
@@ -1,8 +1,0 @@
-// TODO real impl
-export async function login(email: string, password: string) {
-  await new Promise((resolve) => resolve(null))
-  if (false) {
-    return false
-  }
-  return 1
-}
